# api/openapi.yaml
openapi: 3.1.0
info:
  title: IOE Backend API
  version: 0.1.0
  description: >
    Single-tenant LMS backend. Authentication exchanges a Google ID token for an
    access token (Bearer, 15 minutes) and a rotating refresh token held in an
    HttpOnly cookie scoped to /v1/auth.
servers:
  - url: https://api.example.com
paths:
...
  /v1/courses/{courseID}/media/uploads:
    post:
      summary: Start a direct upload of a video or image for a course (owner or root admin)
      description: Rate-limited per user. Upload the bytes to the returned URLs, then complete.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMediaUploadRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaUpload" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/media/uploads/{assetID}/parts:
    post:
      summary: Refresh presigned multipart URLs
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/AssetID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [part_numbers]
              properties:
                part_numbers:
                  type: array
                  minItems: 1
                  maxItems: 1000
                  items: { type: integer, minimum: 1, maximum: 10000 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [part_urls]
                properties:
                  part_urls: { type: array, items: { $ref: "#/components/schemas/MediaUploadPart" } }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/media/uploads/{assetID}/complete:
    post:
      summary: Finish an upload and start processing
      description: Send {} for a single-PUT image upload.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/AssetID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                parts:
                  type: array
                  items:
                    type: object
                    additionalProperties: false
                    required: [part_number, etag]
                    properties:
                      part_number: { type: integer, minimum: 1, maximum: 10000 }
                      etag: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaAsset" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/media/assets/{assetID}:
    get:
      summary: Read an asset's processing state (owner or root admin)
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/AssetID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaAsset" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      summary: Delete an asset; blocks that still reference it resolve to not_found
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/AssetID"
      responses:
        "204": { description: No Content }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409":
          description: asset_in_use; carries lecture_ids array of referencing lectures
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Problem"
                  - type: object
                    properties:
                      lecture_ids:
                        type: array
                        items: { $ref: "#/components/schemas/ID" }
        "502": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/lectures/{lectureID}/media/{assetID}:
    get:
      summary: Resolve playback or display URLs for an asset the lecture references
      description: >
        Same access rules as reading the lecture's content. Not cacheable.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
        - $ref: "#/components/parameters/LectureID"
        - $ref: "#/components/parameters/AssetID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaPlayback" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: __Secure-ioe_refresh
  parameters:
...
    CourseID:
      name: courseID
      in: path
      required: true
      schema: { $ref: "#/components/schemas/ID" }
...
    LectureID:
      name: lectureID
      in: path
      required: true
      schema: { $ref: "#/components/schemas/ID" }
...
    AssetID:
      name: assetID
      in: path
      required: true
      schema: { type: string, pattern: "^[0-9]+$" }
...
  responses:
...
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    MethodNotAllowed:
      description: Method Not Allowed
      headers:
        Allow:
          description: Allowed HTTP methods for the target resource
          required: true
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InternalError:
      description: Internal Server Error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    ID:
      type: string
      pattern: '^[1-9][0-9]{0,18}$'
      description: Snowflake ID rendered as a decimal string.
...
    CreateMediaUploadRequest:
      type: object
      additionalProperties: false
      required: [kind, content_type, filename, size_bytes]
      properties:
        kind: { type: string, enum: [video, image] }
        content_type: { type: string, description: "video/* for videos; image/jpeg, image/png or image/webp for images" }
        filename: { type: string }
        size_bytes: { type: integer, format: int64, minimum: 1 }
    MediaUploadPart:
      type: object
      required: [part_number, url]
      properties:
        part_number: { type: integer }
        url: { type: string, format: uri }
    MediaUpload:
      type: object
      required: [asset_id, expires_at]
      description: Either upload_url (one PUT) or part_size and part_urls (multipart).
      properties:
        asset_id: { type: string }
        upload_id: { type: string }
        upload_url: { type: string, format: uri }
        part_size: { type: integer, format: int64 }
        part_urls: { type: array, items: { $ref: "#/components/schemas/MediaUploadPart" } }
        expires_at: { type: string, format: date-time }
    MediaAsset:
      type: object
      required: [id, course_id, kind, status, progress_percent, updated_at]
      properties:
        id: { type: string }
        course_id: { type: string }
        kind: { type: string, enum: [video, image] }
        status: { type: string, enum: [uploading, processing, ready, failed, canceled, deleting] }
        progress_percent: { type: integer, minimum: 0, maximum: 100 }
        duration_ms: { type: integer, format: int64 }
        width: { type: integer }
        height: { type: integer }
        error_message: { type: string }
        updated_at: { type: string, format: date-time }
    MediaPlayback:
      type: object
      required: [id, kind, status]
      description: >
        URLs are present only when status is ready. A video carries playback_url (a signed HLS
        manifest) and optionally poster_url; an image carries url. URLs expire at expires_at.
      properties:
        id: { type: string }
        kind: { type: string, enum: [video, image] }
        status: { type: string }
        playback_url: { type: string, format: uri }
        url: { type: string, format: uri }
        poster_url: { type: string, format: uri }
        duration_ms: { type: integer, format: int64 }
        width: { type: integer }
        height: { type: integer }
        expires_at: { type: string, format: date-time }
    Problem:
      type: object
      required: [type, code, title, status]
      properties:
        type: { $ref: "#/components/schemas/ProblemType" }
        code:
          allOf: [{ $ref: "#/components/schemas/ProblemType" }]
          description: Same value as type, for clients that read a code member.
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, description: Request ID }
        details:
          type: object
          description: >
            Present on exam_not_open (opens_at), exam_closed (closes_at), reveal_not_yet
            (reveal_at), and edit_* codes (question_id, option_id when known).
          properties:
            opens_at: { type: string, format: date-time }
            closes_at: { type: string, format: date-time }
            reveal_at: { type: string, format: date-time }
            question_id: { $ref: "#/components/schemas/ID" }
            option_id: { $ref: "#/components/schemas/ID" }
    ProblemType:
      type: string
      enum:
        - invalid_request
        - invalid_token
        - refresh_reuse_detected
        - email_unverified
        - origin_not_allowed
        - not_found
        - method_not_allowed
        - payload_too_large
        - unsupported_media_type
        - rate_limited
        - internal
        - forbidden
        - enrollment_required
        - concurrent_modification
        - course_not_editable
        - invalid_transition
        - revision_conflict
        - lecture_content_revision_required
        - patch_too_large
        - block_set_mismatch
        - order_delete_overlap
        - duplicate_client_block_id
        - invalid_client_block_id
        - unsafe_content
        - duplicate_title
        - empty_course
        - unsupported_currency
        - invalid_input
        - exam_not_open
        - exam_closed
        - open_attempt_exists
        - retakes_not_allowed
        - attempt_submitted
        - attempt_expired
        - reveal_attempt_open
        - reveal_disabled
        - reveal_not_yet
        - asset_in_use
        - invalid_quiz_reference
        - invalid_media_reference
        - payment_required
        - course_free
        - already_enrolled
        - already_purchased
        - payment_unavailable
...