# internal/courseauthoring/adapters/postgres/sqlcgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlcgen

import (
	"encoding/json"
	"time"

	"github.com/jackc/pgx/v5/pgtype"
)

type AssessmentExam struct {
	ID           int64
	CourseID     int64
	CreatedAt    time.Time
	UpdatedAt    time.Time
	HeadRevision int32
	DeletedAt    *time.Time
}

type AssessmentExamAttempt struct {
	ID            int64
	ExamID        int64
	CourseID      int64
	UserID        int64
	StartedAt     time.Time
	SubmittedAt   *time.Time
	Score         pgtype.Int4
	Passed        pgtype.Bool
	AutoSubmitted bool
	Answers       json.RawMessage
	Revision      int32
}

type AssessmentExamRevision struct {
	ExamID           int64
	Revision         int32
	Title            string
	Description      string
	Position         int32
	Status           string
	PassMark         int32
	TimeLimitSeconds pgtype.Int4
	RetakesAllowed   bool
	OpensAt          *time.Time
	ClosesAt         *time.Time
	RevealPolicy     string
	Questions        json.RawMessage
	CreatedBy        int64
	CreatedAt        time.Time
}

type AssessmentExamRevisionRow struct {
	ID               int64
	CourseID         int64
	Revision         int32
	HeadRevision     int32
	DeletedAt        *time.Time
	Title            string
	Description      string
	Position         int32
	Status           string
	PassMark         int32
	TimeLimitSeconds pgtype.Int4
	RetakesAllowed   bool
	OpensAt          *time.Time
	ClosesAt         *time.Time
	RevealPolicy     string
	Questions        json.RawMessage
	CreatedAt        time.Time
	UpdatedAt        time.Time
}

type AssessmentQuiz struct {
	ID           int64
	CourseID     int64
	LectureID    int64
	CreatedAt    time.Time
	UpdatedAt    time.Time
	HeadRevision int32
	DeletedAt    *time.Time
}

type AssessmentQuizAttempt struct {
	ID             int64
	QuizID         int64
	UserID         int64
	Answers        json.RawMessage
	IdempotencyKey pgtype.Text
	SubmittedAt    time.Time
	Revision       int32
}

type AssessmentQuizRevision struct {
	QuizID    int64
	Revision  int32
	Position  int32
	Questions json.RawMessage
	CreatedBy int64
	CreatedAt time.Time
}

type AssessmentQuizRevisionRow struct {
	ID           int64
	CourseID     int64
	LectureID    int64
	Revision     int32
	HeadRevision int32
	DeletedAt    *time.Time
	Position     int32
	Questions    json.RawMessage
	CreatedAt    time.Time
	UpdatedAt    time.Time
}

type CourseauthoringCourse struct {
	ID               int64
	OwnerID          int64
	Title            string
	Description      string
	Level            string
	ThumbnailUrl     string
	Status           string
	PriceAmountMinor int64
	PriceCurrency    string
	Version          int64
	CreatedAt        time.Time
	UpdatedAt        time.Time
	SubmittedAt      *time.Time
	ReviewedAt       *time.Time
	ReviewNote       string
	LastVersion      int32
	LiveVersion      *int32
}

type CourseauthoringCourseReview struct {
	ID        int64
	CourseID  int64
	ActorID   int64
	Decision  string
	Note      string
	CreatedAt time.Time
}

type CourseauthoringCourseSubmittedAssessment struct {
	CourseID     int64
	Kind         string
	AssessmentID int64
	Revision     int32
}

type CourseauthoringCourseVersion struct {
	CourseID         int64
	Number           int32
	Title            string
	Description      string
	Level            string
	ThumbnailUrl     string
	PriceAmountMinor int64
	PriceCurrency    string
	PublishedBy      int64
	PublishedAt      time.Time
}

type CourseauthoringCourseVersionAssessment struct {
	CourseID     int64
	Number       int32
	Kind         string
	AssessmentID int64
	Revision     int32
}

type CourseauthoringCourseVersionBlock struct {
	CourseID      int64
	Number        int32
	LectureID     int64
	ID            int64
	Kind          string
	Position      int32
	ClientBlockID string
	Payload       json.RawMessage
}

type CourseauthoringCourseVersionLecture struct {
	CourseID    int64
	Number      int32
	ID          int64
	SectionID   *int64
	Title       string
	FreePreview bool
	SortOrder   int32
}

type CourseauthoringCourseVersionSection struct {
	CourseID  int64
	Number    int32
	ID        int64
	Title     string
	SortOrder int32
}

type CourseauthoringLecture struct {
	ID              int64
	CourseID        int64
	SectionID       *int64
	Title           string
	FreePreview     bool
	SortOrder       int32
	ContentRevision int64
	CreatedAt       time.Time
	UpdatedAt       time.Time
}

type CourseauthoringLectureBlock struct {
	ID            int64
	LectureID     int64
	CourseID      int64
	Kind          string
	Position      int32
	ClientBlockID string
	Payload       json.RawMessage
	CreatedAt     time.Time
	UpdatedAt     time.Time
}

type CourseauthoringSection struct {
	ID        int64
	CourseID  int64
	Title     string
	SortOrder int32
}

type EnrollmentEnrollment struct {
	ID           int64
	CourseID     int64
	UserID       int64
	Status       string
	CancelReason string
	EnrolledAt   time.Time
	CanceledAt   *time.Time
	Version      int64
}

type IdentityRefreshToken struct {
	ID              int64
	UserID          int64
	FamilyID        int64
	TokenHash       []byte
	FamilyExpiresAt time.Time
	ExpiresAt       time.Time
	UsedAt          *time.Time
	RevokedAt       *time.Time
	CreatedAt       time.Time
	UserAgent       string
	Ip              string
}

type IdentityUser struct {
	ID          int64
	GoogleSub   string
	Email       string
	Name        string
	AvatarUrl   string
	Role        string
	CreatedAt   time.Time
	UpdatedAt   time.Time
	LastLoginAt time.Time
}

type MediaAsset struct {
	ID        int64
	CourseID  int64
	Kind      string
	CreatedBy int64
	CreatedAt time.Time
}

type PaymentPurchase struct {
	ID          int64
	UserID      int64
	CourseID    int64
	AmountMinor int64
	Currency    string
	Gateway     string
	GatewayRef  string
	GatewayTxn  string
	Status      string
	CreatedAt   time.Time
	SettledAt   *time.Time
	GrantedAt   *time.Time
	Version     int64
}

type PlatformOutboxMessage struct {
	Offset        int64
	Uuid          string
	CreatedAt     pgtype.Timestamp
	Payload       []byte
	Metadata      []byte
	TransactionID pgtype.Uint64
}

type PlatformOutboxOffset struct {
	ConsumerGroup              string
	OffsetAcked                pgtype.Int8
	LastProcessedTransactionID pgtype.Uint64
}

type ProgressCourseProgress struct {
	CourseID      int64
	UserID        int64
	LastLectureID int64
	UpdatedAt     time.Time
}

type ProgressLectureProgress struct {
	CourseID   int64
	UserID     int64
	LectureID  int64
	State      string
	PositionMs int64
	UpdatedAt  time.Time
}